WIP: Reclustering Approach

+++ src/clustering.cpp
@@ -40,7 +40,7 @@ namespace orangedb {
         // Sample data from the given data
         printf("Running reclustering on %d vectors with %d centroids\n", n, numCentroids);
         for (int i = n-1; i >= 0; i--) {
-            auto query = data + i * dim;
+            auto query = data + ((size_t)i) * ((size_t)dim);
             printf("Query %d: ", i);
             for (int m = 0; m < 10; m++) {
                 printf("%f ", query[m]);
